refactor: DTO virou Nested Class
