--- src/main/java/davidson/helder/api/service/UsuarioLista.java
@@ -1,56 +1,0 @@
-package davidson.helder.api.service;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import davidson.helder.api.model.Usuario;
-
-public abstract class UsuarioLista {
-
-    private static List<Usuario> listaUsuarios = new ArrayList<>();
-
-    public static void addUsuario(Usuario usuario) {
-        if (usuarioExiste(usuario.getId())) return;
-        listaUsuarios.add(usuario);
-    }
-
-    public static void removeUsuario(int id) {
-        if (!usuarioExiste(id)) return;
-
-        for (Usuario indexUsuario : listaUsuarios) {
-            if (indexUsuario.getId() != id) continue;
-
-            listaUsuarios.remove(indexUsuario);
-            return;
-        }
-    }
-
-    private static boolean usuarioExiste(int id) {
-        for (Usuario indexUsuario : listaUsuarios) {
-            if (indexUsuario.getId() != id) continue;
-
-            return true;
-        }
-        return false;
-    }
-
-    public static List<Usuario> getListaUsuarios() {
-        return listaUsuarios;
-    }
-
-    public static Usuario getUsuario(int id){
-
-        if (!usuarioExiste(id)) return null;
-
-        for (Usuario indexUsuario:
-             listaUsuarios) {
-
-            if (indexUsuario.getId() != id) continue;
-
-            return indexUsuario;
-        }
-
-        return null;
-    }
-
-}
