+++ src/test/java/davidson/helder/api/service/UsuarioTest.java
@@ -10,53 +10,51 @@ import davidson.helder.api.model.Usuario;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class UsuarioTest {
+class UsuarioTest {
 
     @BeforeEach
-    public void limparListaECadastrarUsuario(){
-        List<Usuario> listaUsuario = UsuarioLista.getListaUsuarios();
+    void limparListaECadastrarUsuario() {
+        List<Usuario> listaUsuario = Usuario.Lista.getListaUsuarios();
         listaUsuario.clear();
-        
-        assertEquals(0, listaUsuario.size());
-        
 
-        UsuarioCadastro usuario = new UsuarioCadastro(
-            new Usuario(0, "helder", "123", "email")
+        assertEquals(0, listaUsuario.size());
+
+
+        Usuario.Cadastro usuario = new Usuario.Cadastro(
+                new Usuario(0, "helder", "123", "helder@gmail.com")
         );
 
-        usuario.cadastrarUsuario(usuario);
+        Usuario.Cadastro.cadastrarUsuario(usuario);
         assertEquals(1, listaUsuario.size());
     }
-   
+
     @Test
     @DisplayName("Cadastro de usuário")
-    public void cadastrarUsuario() {
-        UsuarioCadastro usuario = new UsuarioCadastro(
-            new Usuario(1, "joao", "@321", "email")
+    void cadastrarUsuario() {
+        Usuario.Cadastro usuario = new Usuario.Cadastro(
+                new Usuario(1, "joao", "123", "joao@bobo.com")
         );
 
-        usuario.cadastrarUsuario(usuario);
+        Usuario.Cadastro.cadastrarUsuario(usuario);
+        assertEquals(2, Usuario.Lista.getListaUsuarios().size());
 
-        usuario.cadastrarUsuario(usuario);
-        assertEquals(2, UsuarioLista.getListaUsuarios().size());
-
-        usuario.cadastrarUsuario(usuario);
-        assertEquals(2, UsuarioLista.getListaUsuarios().size());
+        Usuario.Cadastro.cadastrarUsuario(usuario);
+        assertEquals(2, Usuario.Lista.getListaUsuarios().size());
     }
 
     @Test
     @DisplayName("Exibir um usuário")
-    public void exibirUmUsuario(){
-        List<Usuario> listaUsuario = UsuarioLista.getListaUsuarios();
+    void exibirUmUsuario() {
+        List<Usuario> listaUsuario = Usuario.Lista.getListaUsuarios();
 
-        final int id = UsuarioLista.getUsuario(0).getId();
-        final String nome = UsuarioLista.getUsuario(0).getNome();
-        final String email = UsuarioLista.getUsuario(0).getEmail();
-        final boolean contaAutenticada = UsuarioLista.getUsuario(0).getContaAutenticada();
+        final int id = Usuario.Lista.getUsuario(0).getId();
+        final String nome = Usuario.Lista.getUsuario(0).getNome();
+        final String email = Usuario.Lista.getUsuario(0).getEmail();
+        final boolean contaAutenticada = Usuario.Lista.getUsuario(0).getContaAutenticada();
 
         assertEquals(0, id);
         assertEquals("helder", nome);
-        assertEquals("email", email);
+        assertEquals("helder@gmail.com", email);
         assertFalse(contaAutenticada);
 
     }
